refactor: Move svgs out of components folder
